requirements.txt
flask>=3.1.1
flask-cors>=6.0.1
gunicorn>=22.0.0
langchain>=0.1.0
langchain-openai>=0.0.6
langchain-pinecone>=0.0.3
openai>=1.0.0
pinecone-client>=3.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
pandas>=2.3.0 